Update: Went for a dumb node for simplicity
Updated the code to make it so the Arduino does not do anything until instructed which makes it easier to do interrupts.
Now, the Arduino will check once a second to see if it has been instructed to read the sensors. If it has been instructed, it reads the sensors and sends the data through the serial port to whatever will read it (Right now, my laptop. Later, a Rasperry Pi).